+++ include/playermodelmanager_api.h
@@ -5,7 +5,7 @@ typedef unsigned long PlayerModelManagerHandle;
 
 // Used for keeping compatibility between versions
 // DO NOT EDIT
-#define PMM_API_VERSION 1UL
+#define PMM_API_VERSION 2UL
 
 #define YAZMT_PMM_MOD_NAME "yazmt_mm_playermodelmanager"
 
@@ -136,6 +136,7 @@ typedef enum {
     PMM_DL_ELEGY_OF_EMPTINESS_SHELL_DEKU,  // Only used if model is Deku
     PMM_DL_ELEGY_OF_EMPTINESS_SHELL_GORON, // Only used if model is Goron
     PMM_DL_ELEGY_OF_EMPTINESS_SHELL_ZORA,  // Only used if model is Zora
+    // PMM_DL_ELEGY_OF_EMPTINESS_SHELL_FIERCE_DEITY also exists
 
     // Strength Upgrades (OoT)
     PMM_DL_BRACELET_LFOREARM,
@@ -316,6 +317,9 @@ typedef enum {
     // First Person Slingshot
     PMM_DL_FPS_SLINGSHOT,
 
+    // Elegy of Emptiness Statue (Fierce Deity)
+    PMM_DL_ELEGY_OF_EMPTINESS_SHELL_FIERCE_DEITY, // Only used if model is Fierce Deity
+
     PMM_DL_MAX
 } PlayerModelManagerDisplayListId;
 
@@ -471,9 +475,42 @@ RECOMP_IMPORT(YAZMT_PMM_MOD_NAME, Gfx *PlayerModelManager_getFormDisplayList(uns
 // Returns true if the model attached to the passed in handle is currently equipped, false otherwise.
 RECOMP_IMPORT(YAZMT_PMM_MOD_NAME, bool PlayerModelManager_isApplied(PlayerModelManagerHandle h));
 
-// Allows the tunic color to be set by other mods.
+// Set tunic color of a specific form.
+RECOMP_IMPORT(YAZMT_PMM_MOD_NAME, void PlayerModelManager_requestOverrideFormTunicColor(PlayerTransformation form, u8 r, u8 g, u8 b, u8 a));
+
+// Set tunic color for all forms. Replaces any form-specific tunic colors.
 RECOMP_IMPORT(YAZMT_PMM_MOD_NAME, void PlayerModelManager_requestOverrideTunicColor(u8 r, u8 g, u8 b, u8 a));
 
+// Allows mods to override display lists used by the models used when no custom model is found.
+// This should be used sparingly, but there are some use cases, like if you are exchanging
+// the first person models for non-human forms.
+//
+// Attempts to override shim DLs will be ignored.
+//
+// This function can only be used during the PlayerModelManager_onRegisterModels event.
+//
+// This returns true on successful override, false otherwise.
+RECOMP_IMPORT(YAZMT_PMM_MOD_NAME, bool PlayerModelManager_overrideVanillaDisplayList(unsigned long apiVersion, PlayerTransformation form, PlayerModelManagerDisplayListId displayListId, Gfx *displayList));
+
+// Helper define for PlayerModelManager_overrideVanillaDisplayList. See PlayerModelManager_overrideVanillaDisplayList description for functionality.
+#define PLAYERMODELMANAGER_OVERRIDE_VANILLA_DISPLAY_LIST(form, displayListId, displayList) PlayerModelManager_overrideVanillaDisplayList(PMM_API_VERSION, form, displayListId, displayList)
+
+// Allows mods to override matrixes used by the models used when no custom matrix is found.
+// This should be used sparingly, but there are some use cases, like if you are exchanging
+// equipment models with PlayerModelManager_overrideVanillaDisplayList.
+//
+// This function can only be used during the PlayerModelManager_onRegisterModels event.
+//
+// This returns true on successful override, false otherwise.
+RECOMP_IMPORT(YAZMT_PMM_MOD_NAME, bool PlayerModelManager_overrideVanillaMatrix(unsigned long apiVersion, PlayerTransformation form, PlayerModelManagerMatrixId displayListId, Mtx *mtx));
+
+// Helper define for PlayerModelManager_overrideVanillaDisplayList. See PlayerModelManager_overrideVanillaDisplayList description for functionality.
+#define PLAYERMODELMANAGER_OVERRIDE_VANILLA_MATRIX(form, matrixId, matrix) PlayerModelManager_overrideVanillaMatrix(PMM_API_VERSION, form, matrixId, matrix)
+
+// Returns true if there is a custom player model applied to the passed in form, false otherwise.
+RECOMP_IMPORT(YAZMT_PMM_MOD_NAME, bool PlayerModelManager_isCustomModelApplied(PlayerTransformation form));
+
+// Helper define for register models event.
 #define PLAYERMODELMANAGER_CALLBACK_REGISTER_MODELS RECOMP_CALLBACK(YAZMT_PMM_MOD_NAME, onRegisterModels)
 
 #endif